refactor: move log scanner to internal/airbyte
